// frontend/src/app/model/chat/message.interface.ts
import { Meta } from "@angular/platform-browser";
import { Ichannel } from "./channel.interface";
import { UserI } from "../user/user.interface";

export interface MessageI {
  id?: number;
  type: number;
  text: string;
  user?: UserI;
  channel: Ichannel;
  created_at?: Date;
  updated_at?: Date;
}

export interface MessagePaginateI {
  items: MessageI[];
  meta: Meta;
}